+++ lib/src/home/view/screens/initial_disclosure_agreement.dart
@@ -15,31 +15,22 @@ class InitialDisclosure extends WebPage<InitialDisclosure> {
   List<Widget> children05(BuildContext context) {
     final _screenSize = screenSize;
     return [
-      Stack(
-        children: [
-          SizedBox(
-            height: _screenSize!.height * 0.45,
-            width: _screenSize.width,
-            child: Image.asset(
-              'assets/images/earthNetworked.jpg',
-              fit: BoxFit.cover,
-            ),
-          ),
-          Column(
-            children: [
-              FloatingQuickAccessBar(screen: this),
-              Iframe(
-                height: '1000',
-                width: '640',
-                src:
-                    'https://docs.google.com/document/d/e/2PACX-1vSUGrouFhcTcJqJ4ItmKU0qY6XyF4DIPf619wrlEOgvj-kubE9gV3ME_8K-o04vZqHm4nVEUQcSHVSc/pub?embedded=true',
-              ),
-              const Hyperlink(Text('Print'),
-                  'https://docs.google.com/document/d/1Z2zdFIv_n1HXT2rkU8n6kAOr144b6nHsGbhkUOO9IA0/edit?usp=sharing'),
-            ],
-          )
-        ],
+      SizedBox(
+        height: _screenSize!.height * 0.45,
+        width: _screenSize.width,
+        child: Image.asset(
+          'assets/images/earthNetworked.jpg',
+          fit: BoxFit.cover,
+        ),
       ),
+      Iframe(
+        height: '1000',
+        width: '640',
+        src:
+            'https://docs.google.com/document/d/e/2PACX-1vSUGrouFhcTcJqJ4ItmKU0qY6XyF4DIPf619wrlEOgvj-kubE9gV3ME_8K-o04vZqHm4nVEUQcSHVSc/pub?embedded=true',
+      ),
+      const Hyperlink(Text('Print'),
+          'https://docs.google.com/document/d/1Z2zdFIv_n1HXT2rkU8n6kAOr144b6nHsGbhkUOO9IA0/edit?usp=sharing')
 //      DestinationHeading(screenSize: _screenSize),
     ];
   }